+++ Runtime/Gaze Interaction/IGazeRaycastInteractable.cs
@@ -1,5 +1,5 @@
 // SPDX-FileCopyrightText: Copyright 2023 Holo Interactive <dev@holoi.com>
-// SPDX-FileContributor: Yuchen Zhang <yuchen@holoi.com>
+// SPDX-FileContributor: Yuchen Zhang <yuchenz27@outlook.com>
 // SPDX-License-Identifier: MIT
 
 namespace HoloInteractive.XR.HoloKit