+++ Runtime/iOS/Gaze Gesture Interaction/IGazeGestureInteractable.cs
@@ -1,14 +1,11 @@
 // SPDX-FileCopyrightText: Copyright 2023 Holo Interactive <dev@holoi.com>
-// SPDX-FileContributor: Yuchen Zhang <yuchen@holoi.com>
+// SPDX-FileContributor: Yuchen Zhang <yuchenz27@outlook.com>
 // SPDX-License-Identifier: MIT
 
 namespace HoloInteractive.XR.HoloKit.iOS
 {
     public interface IGazeGestureInteractable : IGazeRaycastInteractable
     {
-        public void OnGestureSelected()
-        {
-
-        }
+        public void OnGestureSelected();
     }
 }