+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2023.12.16
+
+### Added
+
+- Added Gaze Interaction sample
+- Added Gaze Gesture Interaction sample
+- Added Glowing Orbs sample
+
 ## [0.3.12] - 2023.10.28
 
 ### Added